better benchmarks for topological sort & SCC
Adding a topologicalSort / scc directed/undirected setup function
for the graph allows me to change the benchmarks to only test the functions
they are meant to test (as opposed to rebuilding a graph for every test).
Also, the former topological sort bugfix allows me to run the functions
repeatedly and get the same results.