go/graph: use interface{}, not *interface{} for value
In Go, you should use interface{}. An interface{} contains two things:

 - Type identifier
 - Pointer to actual value

You cannot actually assign to a *interface{} field, you will always get
the error "*interface{} is pointer to interface, not interface."

It also doesn't make sense to take a pointer to a interface{}, as the
interface{} contains a pointer anyway.

You can compare two interface{} with == equality comparison and it will
compare both their type and the actual pointer, so the intended behavior
holds.

Signed-off-by: Christian Stewart <christian@paral.in>